+++ CHANGELOG.rst
@@ -2,7 +2,27 @@
 Changelog
 =========
 
+Version 0.1.4
+===========
+
+- feat: add sphinx auto doc
+
+Version 0.1.3
+===========
+
+- feat: auto create template code
+
+Version 0.1.2
+===========
+
+- fix: change gitignore file
+
+Version 0.1.1
+===========
+
+- feat: add src package module
+
 Version 0.1
 ===========
 
-- Initial commit
+- Init project!